Move TreeNodes tests to a separate file
